Changed Classes into the right test files
